# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pitz Pulse",
    "version": "0.1.0"
  },
  "paths": {
...
    "/solicitudes": {
      "post": {
        "summary": "Create",
        "operationId": "create_solicitudes_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            },
            "description": "OK"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      },
      "get": {
        "summary": "List Requests",
        "operationId": "list_requests_solicitudes_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "categoria",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Categoria"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoria"
            }
          },
          {
            "name": "prioridad",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Prioridad"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prioridad"
            }
          },
          {
            "name": "area_sugerida",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Area"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Sugerida"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "classified",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "needs_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Needs Review"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Area": {
        "type": "string",
        "enum": [
          "backend",
          "frontend",
          "data",
          "devops",
          "producto",
          "digital_transformation"
        ],
        "title": "Area"
      },
      "Categoria": {
        "type": "string",
        "enum": [
          "bug",
          "datos",
          "acceso",
          "automatizacion",
          "consulta",
          "otro"
        ],
        "title": "Categoria"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorBody"
      },
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg"
        ],
        "title": "FieldError"
      },
      "Idioma": {
        "type": "string",
        "enum": [
          "es",
          "pt"
        ],
        "title": "Idioma"
      },
      "Item": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "source_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Area"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "classified",
              "failed"
            ],
            "title": "Status"
          },
          "categoria": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Categoria"
              },
              {
                "type": "null"
              }
            ]
          },
          "prioridad": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Prioridad"
              },
              {
                "type": "null"
              }
            ]
          },
          "area_sugerida": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Area"
              },
              {
                "type": "null"
              }
            ]
          },
          "idioma": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Idioma"
              },
              {
                "type": "null"
              }
            ]
          },
          "resumen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumen"
          },
          "requiere_info": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiere Info"
          },
          "pregunta_seguimiento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pregunta Seguimiento"
          },
          "confianza": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confianza"
          },
          "version_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Prompt"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review"
          },
          "corrected": {
            "type": "boolean",
            "title": "Corrected"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "possible_duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Possible Duplicate Of"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message",
          "source_area",
          "status",
          "categoria",
          "prioridad",
          "area_sugerida",
          "idioma",
          "resumen",
          "requiere_info",
          "pregunta_seguimiento",
          "confianza",
          "version_prompt",
          "provider",
          "model",
          "needs_review",
          "corrected",
          "error",
          "created_at",
          "updated_at",
          "possible_duplicate_of"
        ],
        "title": "Item"
      },
...
      "Page": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page"
      },
...
      "Prioridad": {
        "type": "string",
        "enum": [
          "alta",
          "media",
          "baja"
        ],
        "title": "Prioridad"
      },
      "RequestInput": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "source_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Area"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "message"
        ],
        "title": "RequestInput"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}